requirements.txt
jupyter
numpy
pandas
python-binance
mplfinance